[Feature Request] 减少XMOJ服务器负担 Fixes #403
