+++ Server/Source/Process.ts
@@ -1,14 +1,215 @@
 import { Result, ThrowErrorIfFailed } from "./Result";
 import { Database } from "./Database";
-import { Security } from "./Security";
 import { Output } from "./Output";
+import { CaptchaSecretKey } from "./Secret"
+import * as cheerio from "cheerio";
 
 export class Process {
+    private AdminUserList: Array<string> = ["chenlangning", "zhuchenrui2", "shanwenxiao", "admin"];
+    private Username: string;
+    private SessionID: string;
+    private RemoteIP: string;
     private XMOJDatabase: Database;
     private RequestData: Request;
-    private SecurityChecker: Security = new Security();
+    private Fetch = async (RequestURL: URL): Promise<Response> => {
+        Output.Log("Fetch: " + RequestURL.toString());
+        let Abort = new AbortController();
+        setTimeout(() => {
+            Abort.abort();
+        }, 5000);
+        let RequestData = new Request(RequestURL, {
+            headers: {
+                "Cookie": "PHPSESSID=" + this.SessionID
+            },
+            signal: Abort.signal
+        });
+        return await fetch(RequestData);
+    }
+    public CheckParams = (Data: object, Checklist: object): Result => {
+        for (let i in Data) {
+            if (Checklist[i] === undefined) {
+                return new Result(false, "参数" + i + "未知");
+            }
+            const AvailableTypes = ["string", "number", "bigint", "boolean", "symbol", "undefined", "object", "function"];
+            if (AvailableTypes.indexOf(Checklist[i]) === -1) {
+                return new Result(false, "参数类型" + Checklist[i] + "未知");
+            }
+            if (typeof Data[i] !== Checklist[i]) {
+                return new Result(false, "参数" + i + "期望类型" + Checklist[i] + "实际类型" + typeof Data[i]);
+            }
+        }
+        for (let i in Checklist) {
+            if (Data[i] === undefined) {
+                return new Result(false, "参数" + i + "未找到");
+            }
+        }
+        return new Result(true, "参数检测通过");
+    }
+    public CheckToken = async (Data: object): Promise<Result> => {
+        ThrowErrorIfFailed(this.CheckParams(Data, {
+            "SessionID": "string",
+            "Username": "string"
+        }));
+        this.SessionID = Data["SessionID"];
+        this.Username = Data["Username"];
+        // return new Result(true, "令牌检测跳过");
+
+        let CurrentSessionData = ThrowErrorIfFailed(await this.XMOJDatabase.Select("phpsessid", ["user_id", "create_time"], {
+            token: this.SessionID
+        }));
+        if (CurrentSessionData.toString() !== "") {
+            if (CurrentSessionData[0]["user_id"] === this.Username &&
+                CurrentSessionData[0]["create_time"] + 1000 * 60 * 60 * 24 * 7 > new Date().getTime()) {
+                return new Result(true, "令牌匹配");
+            }
+            else {
+                ThrowErrorIfFailed(await this.XMOJDatabase.Delete("phpsessid", {
+                    token: this.SessionID
+                }));
+                Output.Log("Session " + this.SessionID + " expired");
+            }
+        }
+
+        let SessionUsername: string = await this.Fetch(new URL("http://www.xmoj.tech/template/bs3/profile.php"))
+            .then((Response) => {
+                return Response.text();
+            }).then((Response) => {
+                let SessionUsername = Response.substring(Response.indexOf("user_id=") + 8);
+                SessionUsername = SessionUsername.substring(0, SessionUsername.indexOf("'"));
+                return SessionUsername;
+            }).catch((Error) => {
+                Output.Error("Check token failed: " + Error + "\n" +
+                    "PHPSessionID: \"" + this.SessionID + "\"\n" +
+                    "Username    : \"" + this.Username + "\"\n");
+                return "";
+            });
+        if (SessionUsername == "") {
+            Output.Debug("Check token failed: Session invalid\n" +
+                "PHPSessionID: \"" + this.SessionID + "\"\n");
+            return new Result(false, "令牌不合法");
+        }
+        if (SessionUsername != this.Username) {
+            Output.Debug("Check token failed: Session and username not match \n" +
+                "PHPSessionID   : \"" + this.SessionID + "\"\n" +
+                "SessionUsername: \"" + SessionUsername + "\"\n" +
+                "Username       : \"" + this.Username + "\"\n");
+            return new Result(false, "令牌不匹配");
+        }
+
+        ThrowErrorIfFailed(await this.XMOJDatabase.Insert("phpsessid", {
+            token: this.SessionID,
+            user_id: this.Username,
+            create_time: new Date().getTime()
+        }));
+        Output.Log("Record session: " + this.SessionID + " for " + this.Username);
+
+        return new Result(true, "令牌匹配");
+    }
+    public IfUserExist = async (Username: string): Promise<Result> => {
+        return await this.Fetch(new URL("http://www.xmoj.tech/userinfo.php?user=" + Username))
+            .then((Response) => {
+                return Response.text();
+            }).then((Response) => {
+                return new Result(true, "用户检查成功", {
+                    "Exist": Response.indexOf("No such User!") === -1
+                });
+            }).catch((Error) => {
+                Output.Error("Check user exist failed: " + Error + "\n" +
+                    "Username: \"" + Username + "\"\n");
+                return new Result(false, "用户检查失败");
+            });
+    }
+    public IsAdmin = (): boolean => {
+        return this.AdminUserList.indexOf(this.Username) !== -1;
+    }
+    public VerifyCaptcha = async (CaptchaToken: string): Promise<Result> => {
+        const ErrorDescriptions: Object = {
+            "missing-input-secret": "密钥为空",
+            "invalid-input-secret": "密钥不正确",
+            "missing-input-response": "验证码令牌为空",
+            "invalid-input-response": "验证码令牌不正确或已过期",
+            "invalid-widget-id": "解析出的组件编号不正确",
+            "invalid-parsed-secret": "解析出的密钥不正确",
+            "bad-request": "请求格式错误",
+            "timeout-or-duplicate": "相同验证码已经校验过",
+            "internal-error": "服务器错误"
+        };
+        // return new Result(true, "验证码检测跳过");
+        if (CaptchaToken === "") {
+            return new Result(false, "验证码没有完成");
+        }
+        let VerifyFormData = new FormData();
+        VerifyFormData.append("secret", CaptchaSecretKey);
+        VerifyFormData.append("response", CaptchaToken);
+        VerifyFormData.append("remoteip", this.RemoteIP);
+        const VerifyResult = await fetch("https://challenges.cloudflare.com/turnstile/v0/siteverify", {
+            body: JSON.stringify({
+                secret: CaptchaSecretKey,
+                response: CaptchaToken,
+                remoteip: this.RemoteIP
+            }),
+            headers: {
+                "Content-Type": "application/json"
+            },
+            method: 'POST',
+        }).then((Response) => {
+            return Response.json();
+        });
+        if (VerifyResult["success"]) {
+            return new Result(true, "验证码通过");
+        }
+        else {
+            let ErrorString: string = "验证没有通过：";
+            for (let i = 0; i < VerifyResult["error-codes"].length; i++) {
+                ErrorString += (ErrorDescriptions[VerifyResult["error-codes"][i]] == null ? VerifyResult["error-codes"][i] : ErrorDescriptions[VerifyResult["error-codes"][i]]) + " ";
+            }
+            ErrorString = ErrorString.trimEnd();
+            return new Result(false, ErrorString);
+        }
+    }
+    public GetProblemScore = async (ProblemID: number): Promise<number> => {
+        return await this.Fetch(new URL("http://www.xmoj.tech/status.php?user_id=" + this.Username + "&problem_id=" + ProblemID))
+            .then((Response) => {
+                return Response.text();
+            }).then((Response) => {
+                let ParsedDocument: cheerio.CheerioAPI = cheerio.load(Response);
+                let ResultTable = ParsedDocument("#result-tab");
+                if (ResultTable.length == 0) {
+                    Output.Error("Get problem score failed: Cannot find table element\n" +
+                        "ProblemID: \"" + ProblemID + "\"\n" +
+                        "Username : \"" + this.Username + "\"\n");
+                    ThrowErrorIfFailed(new Result(false, "获取题目分数失败"));
+                }
+                let MaxScore: number = 0;
+                let ResultTableBody = ResultTable.children().eq(1);
+                for (let i = 0; i < ResultTableBody.children().length; i++) {
+                    let ResultRow = ResultTableBody.children().eq(i);
+                    if (ResultRow.children().eq(4).text().trim() === "正确") {
+                        return 100;
+                    }
+                    else if (ResultRow.children().eq(4).children().length == 2) {
+                        let ScoreSpan = ResultRow.children().eq(4).children().eq(1);
+                        if (ScoreSpan.length == 0) {
+                            Output.Error("Get problem score failed: Cannot find score span\n" +
+                                "ProblemID: \"" + ProblemID + "\"\n" +
+                                "Username : \"" + this.Username + "\"\n");
+                            ThrowErrorIfFailed(new Result(false, "获取题目分数失败"));
+                        }
+                        let Score: string = ScoreSpan.text().trim();
+                        MaxScore = Math.max(MaxScore, parseInt(Score.substring(0, Score.length - 1)));
+                    }
+                }
+                return MaxScore;
+            }).catch((Error) => {
+                Output.Error("Get user score failed: " + Error + "\n" +
+                    "ProblemID: \"" + ProblemID + "\"\n" +
+                    "Username : \"" + this.Username + "\"\n");
+                ThrowErrorIfFailed(new Result(false, "获取题目分数失败"));
+                return 0;
+            });
+    }
     private AddBBSMention = async (ToUserID: string, PostID: number): Promise<void> => {
-        if (ToUserID === this.SecurityChecker.GetUsername()) {
+        if (ToUserID === this.Username) {
             return;
         }
         if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_mention", {
@@ -52,21 +253,21 @@ export class Process {
     };
     private ProcessFunctions = {
         NewPost: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ProblemID": "number",
                 "Title": "string",
                 "Content": "string",
                 "CaptchaSecretKey": "string"
             }));
-            ThrowErrorIfFailed(await this.SecurityChecker.VerifyCaptcha(Data["CaptchaSecretKey"]));
+            ThrowErrorIfFailed(await this.VerifyCaptcha(Data["CaptchaSecretKey"]));
             let PostID = ThrowErrorIfFailed(await this.XMOJDatabase.Insert("bbs_post", {
-                user_id: this.SecurityChecker.GetUsername(),
+                user_id: this.Username,
                 problem_id: Data["ProblemID"],
                 title: Data["Title"],
                 post_time: new Date().getTime()
             }))["InsertID"];
             let ReplyID = ThrowErrorIfFailed(await this.XMOJDatabase.Insert("bbs_reply", {
-                user_id: this.SecurityChecker.GetUsername(),
+                user_id: this.Username,
                 post_id: PostID,
                 content: Data["Content"],
                 reply_time: new Date().getTime()
@@ -77,12 +278,12 @@ export class Process {
             });
         },
         NewReply: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "PostID": "number",
                 "Content": "string",
                 "CaptchaSecretKey": "string"
             }));
-            ThrowErrorIfFailed(await this.SecurityChecker.VerifyCaptcha(Data["CaptchaSecretKey"]));
+            ThrowErrorIfFailed(await this.VerifyCaptcha(Data["CaptchaSecretKey"]));
 
             let Post = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_post", ["title", "user_id"], { post_id: Data["PostID"] }));
             if (Post.toString() == "") {
@@ -96,7 +297,7 @@ export class Process {
             let MentionPeople = new Array<string>();
             let StringToReplace = new Array<string>();
             for (let Match of String(Data["Content"]).matchAll(/@([a-zA-Z0-9]+)/g)) {
-                if (ThrowErrorIfFailed(await this.SecurityChecker.IfUserExist(Match[1]))["Exist"]) {
+                if (ThrowErrorIfFailed(await this.IfUserExist(Match[1]))["Exist"]) {
                     MentionPeople.push(Match[1]);
                     StringToReplace.push(" <a class=\"link-info\" href=\"http://www.xmoj.tech/userinfo.php?user=" + Match[1] + "\">@" + Match[1] + "</a> ");
                 }
@@ -109,7 +310,7 @@ export class Process {
             });
             MentionPeople = Array.from(new Set(MentionPeople));
             let ReplyID = ThrowErrorIfFailed(await this.XMOJDatabase.Insert("bbs_reply", {
-                user_id: this.SecurityChecker.GetUsername(),
+                user_id: this.Username,
                 post_id: Data["PostID"],
                 content: Data["Content"],
                 reply_time: new Date().getTime()
@@ -119,7 +320,7 @@ export class Process {
                 await this.AddBBSMention(MentionPeople[i], Data["PostID"]);
             }
 
-            if (Post[0]["user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (Post[0]["user_id"] !== this.Username) {
                 await this.AddBBSMention(Post[0]["user_id"], Data["PostID"]);
             }
 
@@ -128,7 +329,7 @@ export class Process {
             });
         },
         GetPosts: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ProblemID": "number",
                 "Page": "number"
             }));
@@ -192,7 +393,7 @@ export class Process {
             return new Result(true, "获得讨论列表成功", ResponseData);
         },
         GetPost: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "PostID": "number",
                 "Page": "number"
             }));
@@ -254,7 +455,7 @@ export class Process {
             return new Result(true, "获得讨论成功", ResponseData);
         },
         LockPost: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "PostID": "number"
             }));
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_post", {
@@ -262,7 +463,7 @@ export class Process {
             }))["TableSize"] === 0) {
                 return new Result(false, "未找到讨论");
             }
-            if (!this.SecurityChecker.IsAdmin()) {
+            if (!this.IsAdmin()) {
                 return new Result(false, "没有权限锁定此讨论");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
@@ -272,13 +473,13 @@ export class Process {
             }
             ThrowErrorIfFailed(await this.XMOJDatabase.Insert("bbs_lock", {
                 post_id: Data["PostID"],
-                lock_person: this.SecurityChecker.GetUsername(),
+                lock_person: this.Username,
                 lock_time: new Date().getTime()
             }));
             return new Result(true, "讨论锁定成功");
         },
         UnlockPost: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "PostID": "number"
             }));
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_post", {
@@ -286,7 +487,7 @@ export class Process {
             }))["TableSize"] === 0) {
                 return new Result(false, "未找到讨论");
             }
-            if (!this.SecurityChecker.IsAdmin()) {
+            if (!this.IsAdmin()) {
                 return new Result(false, "没有权限锁定此讨论");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
@@ -300,7 +501,7 @@ export class Process {
             return new Result(true, "讨论解锁成功");
         },
         EditReply: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ReplyID": "number",
                 "Content": "string"
             }));
@@ -310,7 +511,7 @@ export class Process {
             if (Reply.toString() === "") {
                 return new Result(false, "未找到回复");
             }
-            if (!this.SecurityChecker.IsAdmin() && Reply[0]["user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (!this.IsAdmin() && Reply[0]["user_id"] !== this.Username) {
                 return new Result(false, "没有权限编辑此回复");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_post", {
@@ -319,7 +520,7 @@ export class Process {
                 return new Result(false, "未找到讨论");
             }
 
-            if (!this.SecurityChecker.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
+            if (!this.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
                 post_id: Reply[0]["post_id"]
             }))["TableSize"] === 1) {
                 return new Result(false, "讨论已被锁定");
@@ -332,7 +533,7 @@ export class Process {
             let MentionPeople = new Array<string>();
             let StringToReplace = new Array<string>();
             for (let Match of String(Data["Content"]).matchAll(/@([a-zA-Z0-9]+)/g)) {
-                if (ThrowErrorIfFailed(await this.SecurityChecker.IfUserExist(Match[1]))["Exist"]) {
+                if (ThrowErrorIfFailed(await this.IfUserExist(Match[1]))["Exist"]) {
                     MentionPeople.push(Match[1]);
                     StringToReplace.push(" <a class=\"link-info\" href=\"http://www.xmoj.tech/userinfo.php?user=" + Match[1] + "\">@" + Match[1] + "</a> ");
                 }
@@ -355,7 +556,7 @@ export class Process {
             return new Result(true, "编辑回复成功");
         },
         DeletePost: async (Data: object, CheckUserID: boolean = true): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "PostID": "number"
             }));
             let Post = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_post", ["user_id"], {
@@ -364,12 +565,12 @@ export class Process {
             if (Post.toString() === "") {
                 return new Result(false, "未找到讨论");
             }
-            if (!this.SecurityChecker.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
+            if (!this.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
                 post_id: Data["PostID"]
             }))["TableSize"] === 1) {
                 return new Result(false, "讨论已被锁定");
             }
-            if (!this.SecurityChecker.IsAdmin() && CheckUserID && Post[0]["user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (!this.IsAdmin() && CheckUserID && Post[0]["user_id"] !== this.Username) {
                 return new Result(false, "没有权限删除此讨论");
             }
             let Replies = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_reply", ["reply_id"], {
@@ -384,19 +585,19 @@ export class Process {
             return new Result(true, "删除讨论成功");
         },
         DeleteReply: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ReplyID": "number"
             }));
             let Reply = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_reply", ["user_id", "post_id"], { reply_id: Data["ReplyID"] }));
             if (Reply.toString() === "") {
                 return new Result(false, "未找到回复");
             }
-            if (!this.SecurityChecker.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
+            if (!this.IsAdmin() && ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_lock", {
                 post_id: Reply[0]["post_id"]
             }))["TableSize"] === 1) {
                 return new Result(false, "讨论已被锁定");
             }
-            if (!this.SecurityChecker.IsAdmin() && Reply[0]["user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (!this.IsAdmin() && Reply[0]["user_id"] !== this.Username) {
                 return new Result(false, "没有权限删除此回复");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("bbs_reply", {
@@ -408,12 +609,12 @@ export class Process {
             return new Result(true, "删除回复成功");
         },
         GetBBSMentionList: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {}));
+            ThrowErrorIfFailed(this.CheckParams(Data, {}));
             let ResponseData = {
                 MentionList: new Array<Object>()
             };
             let Mentions = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_mention", ["bbs_mention_id", "post_id", "bbs_mention_time"], {
-                to_user_id: this.SecurityChecker.GetUsername()
+                to_user_id: this.Username
             }));
             for (let i in Mentions) {
                 let Mention = Mentions[i];
@@ -431,12 +632,12 @@ export class Process {
             return new Result(true, "获得讨论提及列表成功", ResponseData);
         },
         GetMailMentionList: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {}));
+            ThrowErrorIfFailed(this.CheckParams(Data, {}));
             let ResponseData = {
                 MentionList: new Array<Object>()
             };
             let Mentions = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message_mention", ["mail_mention_id", "from_user_id", "mail_mention_time"], {
-                to_user_id: this.SecurityChecker.GetUsername()
+                to_user_id: this.Username
             }));
             for (let i in Mentions) {
                 let Mention = Mentions[i];
@@ -449,7 +650,7 @@ export class Process {
             return new Result(true, "获得短消息提及列表成功", ResponseData);
         },
         ReadBBSMention: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "MentionID": "number"
             }));
             let MentionData = ThrowErrorIfFailed(await this.XMOJDatabase.Select("bbs_mention", ["to_user_id"], {
@@ -458,7 +659,7 @@ export class Process {
             if (MentionData.toString() === "") {
                 return new Result(false, "未找到提及");
             }
-            if (MentionData[0]["to_user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (MentionData[0]["to_user_id"] !== this.Username) {
                 return new Result(false, "没有权限阅读此提及");
             }
             ThrowErrorIfFailed(await this.XMOJDatabase.Delete("bbs_mention", {
@@ -467,7 +668,7 @@ export class Process {
             return new Result(true, "阅读讨论提及成功");
         },
         ReadMailMention: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "MentionID": "number"
             }));
             let MentionData = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message_mention", ["to_user_id"], {
@@ -476,7 +677,7 @@ export class Process {
             if (MentionData.toString() === "") {
                 return new Result(false, "未找到提及");
             }
-            if (MentionData[0]["to_user_id"] !== this.SecurityChecker.GetUsername()) {
+            if (MentionData[0]["to_user_id"] !== this.Username) {
                 return new Result(false, "没有权限阅读此提及");
             }
             ThrowErrorIfFailed(await this.XMOJDatabase.Delete("short_message_mention", {
@@ -485,16 +686,16 @@ export class Process {
             return new Result(true, "阅读短消息提及成功");
         },
         GetMailList: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {}));
+            ThrowErrorIfFailed(this.CheckParams(Data, {}));
             let ResponseData = {
                 MailList: new Array<Object>()
             };
             let OtherUsernameList = new Array<string>();
-            let Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["message_from"], { message_to: this.SecurityChecker.GetUsername() }, {}, true));
+            let Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["message_from"], { message_to: this.Username }, {}, true));
             for (let i in Mails) {
                 OtherUsernameList.push(Mails[i]["message_from"]);
             }
-            Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["message_to"], { message_from: this.SecurityChecker.GetUsername() }, {}, true));
+            Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["message_to"], { message_from: this.Username }, {}, true));
             for (let i in Mails) {
                 OtherUsernameList.push(Mails[i]["message_to"]);
             }
@@ -502,14 +703,14 @@ export class Process {
             for (let i in OtherUsernameList) {
                 let LastMessageFrom = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["content", "send_time"], {
                     message_from: OtherUsernameList[i],
-                    message_to: this.SecurityChecker.GetUsername()
+                    message_to: this.Username
                 }, {
                     Order: "send_time",
                     OrderIncreasing: false,
                     Limit: 1
                 }));
                 let LastMessageTo = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", ["content", "send_time"], {
-                    message_from: this.SecurityChecker.GetUsername(),
+                    message_from: this.Username,
                     message_to: OtherUsernameList[i]
                 }, {
                     Order: "send_time",
@@ -528,7 +729,7 @@ export class Process {
                 }
                 let UnreadCount = ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("short_message", {
                     message_from: OtherUsernameList[i],
-                    message_to: this.SecurityChecker.GetUsername(),
+                    message_to: this.Username,
                     is_read: 0
                 }));
                 ResponseData.MailList.push({
@@ -544,29 +745,29 @@ export class Process {
             return new Result(true, "获得短消息列表成功", ResponseData);
         },
         SendMail: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ToUser": "string",
                 "Content": "string"
             }));
-            if (ThrowErrorIfFailed(await this.SecurityChecker.IfUserExist(Data["ToUser"]))["Exist"] === false) {
+            if (ThrowErrorIfFailed(await this.IfUserExist(Data["ToUser"]))["Exist"] === false) {
                 return new Result(false, "未找到用户");
             }
-            if (Data["ToUser"] === this.SecurityChecker.GetUsername()) {
+            if (Data["ToUser"] === this.Username) {
                 return new Result(false, "无法给自己发送短消息");
             }
             let MessageID = ThrowErrorIfFailed(await this.XMOJDatabase.Insert("short_message", {
-                message_from: this.SecurityChecker.GetUsername(),
+                message_from: this.Username,
                 message_to: Data["ToUser"],
                 content: Data["Content"],
                 send_time: new Date().getTime()
             }))["InsertID"];
-            await this.AddMailMention(this.SecurityChecker.GetUsername(), Data["ToUser"]);
+            await this.AddMailMention(this.Username, Data["ToUser"]);
             return new Result(true, "发送短消息成功", {
                 MessageID: MessageID
             });
         },
         GetMail: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "OtherUser": "string"
             }));
             let ResponseData = {
@@ -574,7 +775,7 @@ export class Process {
             };
             let Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", [], {
                 message_from: Data["OtherUser"],
-                message_to: this.SecurityChecker.GetUsername()
+                message_to: this.Username
             }, {
                 Order: "send_time",
                 OrderIncreasing: false
@@ -591,7 +792,7 @@ export class Process {
                 });
             }
             Mails = ThrowErrorIfFailed(await this.XMOJDatabase.Select("short_message", [], {
-                message_from: this.SecurityChecker.GetUsername(),
+                message_from: this.Username,
                 message_to: Data["OtherUser"]
             }, {
                 Order: "send_time",
@@ -615,30 +816,84 @@ export class Process {
                 is_read: 1
             }, {
                 message_from: Data["OtherUser"],
-                message_to: this.SecurityChecker.GetUsername()
+                message_to: this.Username
             });
             return new Result(true, "获得短消息成功", ResponseData);
         },
         UploadStd: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ProblemID": "number"
             }));
-            if (await this.SecurityChecker.GetProblemScore(Data["ProblemID"]) !== 100) {
+            let ProblemID = Data["ProblemID"];
+            if (await this.GetProblemScore(ProblemID) !== 100) {
                 return new Result(false, "没有权限上传此标程");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("std_answer", {
-                problem_id: Data["ProblemID"]
+                problem_id: ProblemID
             }))["TableSize"] !== 0) {
                 return new Result(true, "此题已经有人上传标程");
             }
+            var StdCode: string = "";
+            var PageIndex: number = 0;
+            while (StdCode === "") {
+                await this.Fetch(new URL("http://www.xmoj.tech/problemstatus.php?id=" + ProblemID + "&page=" + PageIndex))
+                    .then((Response) => {
+                        return Response.text();
+                    }).then(async (Response) => {
+                        if (Response.indexOf("[NEXT]") === -1) {
+                            StdCode = "这道题没有标程（即用户std没有AC这道题）";
+                            return;
+                        }
+                        let ParsedDocument: cheerio.CheerioAPI = cheerio.load(Response);
+                        let SubmitTable = ParsedDocument("#problemstatus");
+                        if (SubmitTable.length == 0) {
+                            Output.Error("Get Std code failed: Cannot find submit table\n" +
+                                "ProblemID: \"" + ProblemID + "\"\n" +
+                                "Username : \"" + this.Username + "\"\n");
+                            ThrowErrorIfFailed(new Result(false, "获取标程失败"));
+                        }
+                        let SubmitTableBody = SubmitTable.children().eq(1);
+                        for (let i = 1; i < SubmitTableBody.children().length; i++) {
+                            let SubmitRow = SubmitTableBody.children().eq(i);
+                            if (SubmitRow.children().eq(2).text().trim() === "std") {
+                                let SID: string = SubmitRow.children().eq(1).text();
+                                if (SID.indexOf("(") != -1) {
+                                    SID = SID.substring(0, SID.indexOf("("));
+                                }
+                                await this.Fetch(new URL("http://www.xmoj.tech/getsource.php?id=" + SID))
+                                    .then((Response) => {
+                                        return Response.text();
+                                    })
+                                    .then((Response) => {
+                                        Response = Response.substring(0, Response.indexOf("<!--not cached-->")).trim();
+                                        if (Response === "I am sorry, You could not view this code!") {
+                                            Output.Error("Get Std code failed: Cannot view code\n" +
+                                                "ProblemID: \"" + ProblemID + "\"\n" +
+                                                "Username : \"" + this.Username + "\"\n");
+                                            ThrowErrorIfFailed(new Result(false, "获取标程失败"));
+                                        }
+                                        Response = Response.substring(0, Response.indexOf("/**************************************************************")).trim();
+                                        StdCode = Response;
+                                    });
+                            }
+                        }
+                    }).catch((Error) => {
+                        Output.Error("Get Std code failed: " + Error + "\n" +
+                            "ProblemID: \"" + ProblemID + "\"\n" +
+                            "Username : \"" + this.Username + "\"\n");
+                        ThrowErrorIfFailed(new Result(false, "获取标程失败"));
+                    });
+                PageIndex++;
+            }
+
             ThrowErrorIfFailed(await this.XMOJDatabase.Insert("std_answer", {
                 problem_id: Data["ProblemID"],
-                std_code: await this.SecurityChecker.GetStdCode(Data["ProblemID"])
+                std_code: StdCode
             }));
             return new Result(true, "标程上传成功");
         },
         GetStdList: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {}));
+            ThrowErrorIfFailed(this.CheckParams(Data, {}));
             let ResponseData = {
                 StdList: new Array<number>()
             };
@@ -649,10 +904,10 @@ export class Process {
             return new Result(true, "获得标程列表成功", ResponseData);
         },
         GetStd: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "ProblemID": "number"
             }));
-            if (await this.SecurityChecker.GetProblemScore(Data["ProblemID"]) < 50) {
+            if (await this.GetProblemScore(Data["ProblemID"]) < 50) {
                 return new Result(false, "没有权限获取此标程");
             }
             let Std = ThrowErrorIfFailed(await this.XMOJDatabase.Select("std_answer", ["std_code"], {
@@ -666,10 +921,10 @@ export class Process {
             });
         },
         NewBadge: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "UserID": "string"
             }));
-            if (!this.SecurityChecker.IsAdmin()) {
+            if (!this.IsAdmin()) {
                 return new Result(false, "没有权限创建此标签");
             }
             ThrowErrorIfFailed(await this.XMOJDatabase.Insert("badge", {
@@ -678,13 +933,13 @@ export class Process {
             return new Result(true, "创建标签成功");
         },
         EditBadge: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "UserID": "string",
                 "BackgroundColor": "string",
                 "Color": "string",
                 "Content": "string"
             }));
-            if (!this.SecurityChecker.IsAdmin() && Data["UserID"] !== this.SecurityChecker.GetUsername()) {
+            if (!this.IsAdmin() && Data["UserID"] !== this.Username) {
                 return new Result(false, "没有权限编辑此标签");
             }
             if (ThrowErrorIfFailed(await this.XMOJDatabase.GetTableSize("badge", {
@@ -702,7 +957,7 @@ export class Process {
             return new Result(true, "编辑标签成功");
         },
         GetBadge: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "UserID": "string"
             }));
             let BadgeData = ThrowErrorIfFailed(await this.XMOJDatabase.Select("badge", ["background_color", "color", "content"], {
@@ -718,10 +973,10 @@ export class Process {
             });
         },
         DeleteBadge: async (Data: object): Promise<Result> => {
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(Data, {
+            ThrowErrorIfFailed(this.CheckParams(Data, {
                 "UserID": "string"
             }));
-            if (!this.SecurityChecker.IsAdmin()) {
+            if (!this.IsAdmin()) {
                 return new Result(false, "没有权限删除此标签");
             }
             ThrowErrorIfFailed(await this.XMOJDatabase.Delete("badge", {
@@ -733,7 +988,7 @@ export class Process {
     constructor(RequestData: Request, Environment) {
         this.XMOJDatabase = new Database(Environment.DB);
         this.RequestData = RequestData;
-        this.SecurityChecker.SetRemoteIP(RequestData.headers.get("CF-Connecting-IP") || "");
+        this.RemoteIP = RequestData.headers.get("CF-Connecting-IP") || "";
     }
     public async Process(): Promise<Result> {
         try {
@@ -756,18 +1011,18 @@ export class Process {
             catch (Error) {
                 throw new Result(false, "请求格式有误");
             }
-            ThrowErrorIfFailed(this.SecurityChecker.CheckParams(RequestJSON, {
+            ThrowErrorIfFailed(this.CheckParams(RequestJSON, {
                 "Authentication": "object",
                 "Data": "object"
             }));
             var TokenFailedCount = 0;
             while (true) {
-                if ((await this.SecurityChecker.CheckToken(RequestJSON["Authentication"])).Data["Success"]) {
+                if ((await this.CheckToken(RequestJSON["Authentication"])).Data["Success"]) {
                     break;
                 }
                 TokenFailedCount++;
                 if (TokenFailedCount >= 2) {
-                    ThrowErrorIfFailed(await this.SecurityChecker.CheckToken(RequestJSON["Authentication"]));
+                    ThrowErrorIfFailed(await this.CheckToken(RequestJSON["Authentication"]));
                     break;
                 }
             }
