+++ Server/Source/index.ts
@@ -11,11 +11,21 @@ export default {
         });
     },
     async scheduled(Event, Environment, Context) {
-        Context.waitUntil(new Database(Environment.DB).Delete("short_message", {
-            "send_time": {
-                "Operator": "<=",
-                "Value": new Date().getTime() - 1000 * 60 * 60 * 24 * 7
-            }
+        let XMOJDatabase = new Database(Environment.DB);
+        Context.waitUntil(new Promise<void>(async (Resolve) => {
+            await XMOJDatabase.Delete("short_message", {
+                "send_time": {
+                    "Operator": "<=",
+                    "Value": new Date().getTime() - 1000 * 60 * 60 * 24 * 7
+                }
+            });
+            await XMOJDatabase.Delete("phpsessid", {
+                "create_time": {
+                    "Operator": "<=",
+                    "Value": new Date().getTime() - 1000 * 60 * 60 * 24 * 7
+                }
+            });
+            Resolve();
         }));
     },
 };
